Updates and fixes to authentication
