+++ src/client/router.ts
@@ -1,21 +1,21 @@
 import VueRouter from "vue-router";
 import { HomePage, LoginPage, ShowClassPage, RiderPage, AdminPage } from "./pages";
-import { state, signOut, post, apiurl } from "./helpers";
+import { state, signOut, post, apiurl, clearSession } from "./helpers";
 
 export const router = new VueRouter({
     linkActiveClass: "is-active",
     routes: [
-        { path: "/", redirect: "/home", meta: { admin: false } },
-        { path: "/home", component: HomePage, meta: { admin: false } },
-        { path: "/login", component: LoginPage, meta: { admin: false } },
-        { path: "/riders", component: RiderPage, meta: { admin: false } },
-        { path: "/class", component: ShowClassPage, meta: { admin: false } },
-        { path: "/admin", component: AdminPage, meta: { admin: false } }
+        { path: "/", redirect: "/home", meta: { anon: true } },
+        { path: "/home", component: HomePage, meta: { anon: true } },
+        { path: "/login", component: LoginPage, meta: { anon: true } },
+        { path: "/riders", component: RiderPage, meta: { anon: true } },
+        { path: "/class", component: ShowClassPage, meta: { anon: true } },
+        { path: "/admin", component: AdminPage, meta: { anon: true } }
     ]
 });
 
 router.beforeEach((to, _from, next) => {
-    if (!state.get().loggedIn && to.meta.admin && to.path !== "/login") {
+    if (!state.get().loggedIn && !to.meta.anon && to.path !== "/login") {
         return next("/login");
     }
 
@@ -25,6 +25,10 @@ router.beforeEach((to, _from, next) => {
 post(`${apiurl}/auth/session-is-valid`, { token: state.get().token })
     .then(sessionState => {
         if (sessionState.isValid) return;
+        if (router.currentRoute.meta.anon) {
+            clearSession();
+            return;
+        }
         return signOut(router);
     })
     .catch(e => console.log(e));
